Fixed local import error in pyright
