+++ classification/models/early_stopper.py
@@ -1,0 +1,36 @@
+""" Contains the early stopper class
+    TODO
+
+"""
+
+import torch
+from torch import nn
+from utils import print_verbose
+import config 
+# timm is a library of premade models
+import timm
+
+cfg = config.cfg
+
+class EarlyStopper():
+    """Stop when the model is no longer improving
+    """
+    def __init__(self, patience=3, min_delta=0):
+        self.patience = patience # epochs to wait before stopping
+        self.min_delta = min_delta # min change that counts as improvement
+        self.counter = 0
+        self.max_valid_map = 0
+
+    def early_stop(self, valid_map):
+        """ Returns True if the model should early stop
+        """
+        # reset counter if it improved by more than min_delta
+        if valid_map > self.max_valid_map + self.min_delta:
+            self.max_valid_map = valid_map
+            self.counter = 0
+        # increase counter if it has not improved
+        elif valid_map < (self.max_valid_map - self.min_delta):
+            self.counter += 1
+            if self.counter >= self.patience:
+                return True
+        return False
