+++ classification/models/timm_model.py
@@ -1,0 +1,55 @@
+# pylint: disable=E1123:
+
+""" Contains the model class
+    Model: model with forward pass method. Generated automatically from a timm model
+
+"""
+import torch
+from torch import nn
+from utils import print_verbose
+import config 
+# timm is a library of premade models
+import timm
+from .model import bce_loss_fn, cross_entropy_loss_fn
+
+cfg = config.cfg
+
+# pylint: disable=too-many-instance-attributes
+class TimmModel(nn.Module):
+    """ Efficient net neural network
+    """
+    # pylint: disable=too-many-arguments
+    def __init__(self,
+                 num_classes,
+                 model_name="tf_efficientnet_b4",
+                 pretrained=True):
+        """ Initializes the model
+        """
+        super().__init__()
+        self.num_classes = num_classes
+        # See config.py for list of recommended models
+        self.model = timm.create_model(model_name, pretrained=pretrained, num_classes=num_classes)
+        self.loss_fn = None
+        self.without_logits = cfg.loss_fnc == "BCE"
+
+        print_verbose("add sigmod: ", self.without_logits, verbose=cfg.verbose)
+    
+    def forward(self, images):
+        """ Forward pass of the model
+        """
+        x = self.model(images)
+        if self.without_logits:
+            x = torch.sigmoid(x)
+        return x
+
+    def create_loss_fn(self,train_dataset):
+        """ Returns the loss function and sets self.loss_fn
+        """
+        loss_desc = cfg.loss_fnc
+        if loss_desc == "CE":
+            return cross_entropy_loss_fn(self, train_dataset)#
+        if loss_desc == "BCE":
+            return bce_loss_fn(self, self.without_logits)
+        if loss_desc == "BCEWL":
+            return bce_loss_fn(self, self.without_logits)
+        raise RuntimeError("Pick a loss in the form of CE, BCE, BCEWL")
